# web/src/shared/api/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Tamandua",
    "version": "0.9"
  },
  "paths": {
...
    "/api/images/scans": {
      "post": {
        "tags": [
          "images"
        ],
        "summary": "Image Scan",
        "description": "Queues the scan of a container image, pulled from its registry.",
        "operationId": "image_scan_api_images_scans_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "properties": {
                  "reference": {
                    "type": "string",
                    "title": "Reference"
                  },
                  "context": {
                    "type": "string",
                    "title": "Context",
                    "default": ""
                  }
                },
                "additionalProperties": false,
                "type": "object",
                "required": [
                  "reference"
                ],
                "title": "ImageScanIn"
              }
            }
          },
          "required": true
        },
        "responses": {
          "202": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/QueuedImageScan"
                }
              }
            }
          }
        }
      }
    },
    "/api/images/batches": {
      "post": {
        "tags": [
          "images"
        ],
        "summary": "Image Batch",
        "description": "Several images at once (up to 100), in a batch that moves on when the server is free.",
        "operationId": "image_batch_api_images_batches_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "properties": {
                  "references": {
                    "items": {
                      "type": "string"
                    },
                    "type": "array",
                    "maxItems": 100,
                    "title": "References"
                  },
                  "context": {
                    "type": "string",
                    "title": "Context",
                    "default": ""
                  }
                },
                "additionalProperties": false,
                "type": "object",
                "required": [
                  "references"
                ],
                "title": "ImageBatchIn",
                "description": "Blank and repeated references are dropped before counting."
              }
            }
          },
          "required": true
        },
        "responses": {
          "202": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BatchSummary"
                }
              }
            }
          }
        }
      }
    },
    "/api/registries": {
      "get": {
        "tags": [
          "images"
        ],
        "summary": "Registry List",
        "description": "Registries with saved credentials (never the token), and whether private addresses are allowed.",
        "operationId": "registry_list_api_registries_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "type": "object",
                  "title": "Response Registry List Api Registries Get"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "images"
        ],
        "summary": "Registry Save",
        "description": "Read-only credentials of a private registry: stored encrypted, never sent back to the browser.",
        "operationId": "registry_save_api_registries_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "properties": {
                  "action": {
                    "type": "string",
                    "enum": [
                      "save",
                      "remove"
                    ],
                    "title": "Action"
                  },
                  "registry": {
                    "type": "string",
                    "maxLength": 200,
                    "title": "Registry"
                  },
                  "username": {
                    "anyOf": [
                      {
                        "type": "string",
                        "maxLength": 200
                      },
                      {
                        "type": "null"
                      }
                    ],
                    "title": "Username"
                  },
                  "token": {
                    "anyOf": [
                      {
                        "type": "string"
                      },
                      {
                        "type": "null"
                      }
                    ],
                    "title": "Token"
                  }
                },
                "additionalProperties": false,
                "type": "object",
                "required": [
                  "action",
                  "registry"
                ],
                "title": "RegistryIn",
                "description": "`save` needs `username` and `token`; `remove` only the registry."
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "type": "object",
                  "title": "Response Registry Save Api Registries Post"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "BatchSummary": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "label": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Label"
          },
          "status": {
            "type": "string",
            "title": "Status"
          },
          "created_at": {
            "type": "string",
            "title": "Created At"
          },
          "by": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "By"
          },
          "total": {
            "type": "integer",
            "title": "Total"
          },
          "pending": {
            "type": "integer",
            "title": "Pending"
          },
          "running": {
            "type": "integer",
            "title": "Running"
          },
          "done": {
            "type": "integer",
            "title": "Done"
          },
          "failed": {
            "type": "integer",
            "title": "Failed"
          },
          "critical": {
            "type": "integer",
            "title": "Critical"
          },
          "high": {
            "type": "integer",
            "title": "High"
          },
          "eta_seconds": {
            "type": "integer",
            "title": "Eta Seconds"
          },
          "failed_items": {
            "items": {
              "additionalProperties": true,
              "type": "object"
            },
            "type": "array",
            "maxItems": 20,
            "title": "Failed Items"
          }
        },
        "type": "object",
        "required": [
          "id",
          "label",
          "status",
          "created_at",
          "by",
          "total",
          "pending",
          "running",
          "done",
          "failed",
          "critical",
          "high",
          "eta_seconds",
          "failed_items"
        ],
        "title": "BatchSummary",
        "description": "Progress worked out from the batch's real runs, with the estimated time left."
      },
...
      "QueuedImageScan": {
        "properties": {
          "run": {
            "additionalProperties": true,
            "type": "object",
            "title": "Run"
          },
          "image": {
            "additionalProperties": true,
            "type": "object",
            "title": "Image"
          }
        },
        "type": "object",
        "required": [
          "run",
          "image"
        ],
        "title": "QueuedImageScan"
      },
...
    },
    "securitySchemes": {
      "session": {
        "type": "apiKey",
        "in": "cookie",
        "name": "tamandua_session",
        "description": "Session signed in to the panel."
      },
      "metrics": {
        "type": "http",
        "scheme": "bearer",
        "description": "TAMANDUA_METRICS_TOKEN (Prometheus scraper)."
      }
    }
  },
  "security": [
    {
      "session": []
    }
  ]
}